fix(compiler): decouple EVM metrics from frontend headers
Keep stack-lift statistics in the lightweight metrics header and let the EVM frontend consume that definition. Avoid pulling EVMC-only headers into generic multipass compiler builds that do not enable EVM support.
